Use half precision for NeuralNetworkGenomeTO weights and add channel copy helper functions

+++ source/EngineGpuKernels/EntityFactory.cuh
@@ -122,7 +122,7 @@ __inline__ __device__ Genome* EntityFactory::createGenomeFromTO(TOs const& to, i
             node.color = nodeTO.color;
             node.numAdditionalConnections = nodeTO.numAdditionalConnections;
             for (int i = 0; i < MAX_CHANNELS * MAX_CHANNELS; ++i) {
-                node.neuralNetwork.weights[i] = nodeTO.neuralNetwork.weights[i];
+                node.neuralNetwork.weights[i] = nodeTO.neuralNetwork.weights[i];  // Both are half, direct copy
             }
             for (int i = 0; i < MAX_CHANNELS; ++i) {
                 node.neuralNetwork.biases[i] = nodeTO.neuralNetwork.biases[i];
@@ -760,7 +760,7 @@ __inline__ __device__ Object* EntityFactory::createCellFromNode(
 
     cell.neuralNetwork = _data->entities.heap.getTypedSubArray<NeuralNetwork>(1);
     for (int i = 0; i < MAX_CHANNELS * MAX_CHANNELS; ++i) {
-        cell.neuralNetwork->weights[i] = __float2half(node->neuralNetwork.weights[i]);
+        cell.neuralNetwork->weights[i] = node->neuralNetwork.weights[i];  // Both are half, direct copy
     }
     for (int i = 0; i < MAX_CHANNELS; ++i) {
         cell.neuralNetwork->biases[i] = node->neuralNetwork.biases[i];
